requirements.txt
python-decouple>=3.0
Django>=4.2.5
djangorestframework>=3.14.0
Pillow>=10.0.0
django-cors-headers>=4.0.0
psycopg2-binary>=2.9.6
gunicorn>=20.1.0
whitenoise>=6.4.0
python-multipart>=0.0.6
qrcode>=7.4.2
djangorestframework-simplejwt>=5.2.2